requirements.txt
arrow==0.16.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
dogpile.cache==1.0.2
idna==2.10
importlib-metadata==2.0.0
pbr==5.5.0
prometheus-client==0.8.0
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
stevedore==3.2.2
urllib3==1.25.10
zipp==3.2.0
